Refactor player duplicate to manage collisions
If the player duplicate goes out of bounds, it needs to be destroyed.

While realizing this has to happen on the player duplicate prefab, I consolidated all collisions here.

As the player has the Tilemap info for movement, we make it public so we can use it.

Added benefit is one less script to initialize in the Obstacle Tilemap.